Refactoring of EbicsUserCertificateManager.kt Added few tests
